UPDATE-robot

+++ src/ros2_ws/src/robot_service/robot_service/robot.py
@@ -37,7 +37,7 @@ class Robot(Node):
         """
         input_text = self.listener_callback()
         match = re.findall(r'\b\d+\b', input_text)
-        position = [int(match) for i in match[-2:]]
+        position = [float(match) for i in match[-2:]]
         return position
 
     def move_towards_required_position(self):
